// src/types/admin.ts
export interface UserSubmission {
  name: string;
  email: string;
  gender: string;
  cpf: string;
  city: string;
  state: string;
  selectedPackage: string;
  installments: string;
  dueDate: string;
  phone: string;
  address: string;
  number: string;
  complement?: string;
  neighborhood: string;
  zipCode: string;
  height: string;
  weight: string;
  age: string;
}